update to v2.4.0

+++ src/daemon/zbdriver/silicon_labs/Simplicity-Studio-v4/developer/sdks/gecko_sdk_suite/v2.7/protocol/zigbee/app/framework/include/af-types.h
@@ -1874,7 +1874,7 @@ typedef struct {
 
 typedef struct {
   EmberNodeId emberNodeId;
-  uint32_t timeStamp;
+  uint64_t timeStamp;
 } EmberAfJoiningDevice;
 
 #define EMBER_AF_INVALID_CLUSTER_ID 0xFFFF
