+++ src/daemon/zbdriver/silicon_labs/Simplicity-Studio-v4/developer/sdks/gecko_sdk_suite/v2.7/protocol/zigbee/app/framework/plugin/device-table/device-table.h
@@ -34,7 +34,7 @@ enum {
 typedef struct {
   uint8_t     endpoint;
   uint16_t    deviceId;
-  uint32_t    lastMsgTimestamp;
+  uint64_t    lastMsgTimestamp;
   EmberNodeId nodeId;
   EmberEUI64  eui64;
   EmberAfPluginDeviceTableDeviceState state;
