+++ src/daemon/zbdriver/silicon_labs/Simplicity-Studio-v4/developer/sdks/gecko_sdk_suite/v2.7/protocol/zigbee/app/framework/plugin/reporting/reporting.h
@@ -27,7 +27,7 @@
 #endif
 
 typedef struct {
-  uint32_t lastReportTimeMs;
+  uint64_t lastReportTimeMs;
   EmberAfDifferenceType lastReportValue;
   bool reportableChange;
 } EmAfPluginReportVolatileData;
